requirements.txt
# Core dependencies
pydantic==2.5.0
click==8.1.0
PyYAML==6.0.1
python-slugify==8.0.0
python-dotenv==1.0.0
tqdm==4.65.0
colorama==0.4.6

# Image processing
Pillow==10.2.0

# Video/Audio processing
ffmpeg-python==0.2.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
numpy==1.26.0
humanize==4.9.0

# AI/Content generation
openai==1.30.0
tenacity==8.2.0
rich==13.0.0